Up to Gogoat
add movesets and changes for first half of gen 6 pkmn up to Gogoat. Fix Battle Bond to function for ordinary Greninja (as opposed to Battle Bond Greninja form which is unobtainable in the hack for players or opponents). Add Rock Climb to Pokémon who learned it by HM in gen 4, as it is now
a powerful and relevant move.